Added info : tscli scheduled-pinboard enable
tscli scheduled-pinboard enable : needs to set up along with a whitelist of domains. Without that option specified, this one command is somewhat risky and shouldn't be implemented. Without a whitelist it may not even work (as the list of whitelisted domains will be empty).
So I've added a tip about that.
Ideally, the whitelist needs to be specified via tscli as well.