Updated Quick Select section
When you type a phrase or a letter in the search bar, the first suggestion is automatically highlighted and you can press Enter to select it. Suggestions can be an attribute, measures, value, keyword, historical queries, or exactly what you typed as the search phrase. Historic suggestions always display on top with the first suggestion highlighted as default.  If there are no suggestions, the suggestion list is not be displayed.