Skill set guidelines for TS admin role
Bsaed on the query received from Siemens EMEA today, I've added three basic skills (and levels) that would typically be necessary for someone to be an effective TS admin for an organisation.

Please review for inclusion. Check with Siva as well - in case we do not want to publish such information, but I personally see no harm as it helps the customers.