Updated SSL installation doc.
In addition to the changes highlighted in the draft, we also need to make sure that the following text on the page is a clickable URL and not plain text.

" To test that the certificate is correctly installed, sign into the ThoughtSpot application (https://docs.thoughtspot.com/6.0/admin/setup/logins.html#sign-in-to-the-thoughtspot-application)."

It will be worth searching for "https://" string on all docs pages to eliminate such cases. I am seeing more and more of these prop up.